Configure webpack, transform existing html pages into web components and folder restructure
