--- client/pages/index.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html> 
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Sliding Puzzle</title>
-
-    <link rel="stylesheet" type="text/css" href="../styles/style.css">
-    <script src="../scripts/timer.js" type="module"></script>
-    <script src="../scripts/index.js" type="module"></script>
-
-</head>
-    <body>
-        <div><button class="sign-out-btn">Sign out</button> <span id="username"></span></div>
-        <main class="sliding-puzzle-figure">
-            
-            <game-board id="playerA" isready="false"></game-board>
-        
-            <div class="timer">
-                <button class="ranking">Ranking</button>
-                <p><span id="hours">00</span>:<span id="minuets">00</span>:<span id="seconds">00</span>:<span id="tens">00</span></p>
-                <div class="vertical-line"></div>
-            </div>
-
-            <game-board id="playerB" isready="false"></game-board>
-        
-        </main>
-        <div class="chat-box">
-            <label for="chat-box">Chat:</label>
-            <div class="chat"></div>
-        </div>
-    </body>
-</html>
